Extracted function checking state
