+++ semantics/lpg_local_variable.c
@@ -1,6 +1,9 @@
+#include <lpg_source_location.h>
 #include "lpg_local_variable.h"
 #include "lpg_allocate.h"
 #include "lpg_for.h"
+#include "lpg_instruction.h"
+#include "function_checking_state.h"
 
 local_variable local_variable_create(unicode_string name, type const type_, optional_value compile_time_value,
                                      register_id where)
@@ -44,4 +47,52 @@ variable_address variable_address_from_local(register_id const local)
 {
     variable_address const result = {optional_capture_index_empty, local};
     return result;
+}
+
+read_local_variable_result read_local_variable(LPG_NON_NULL(function_checking_state *const state),
+                                               instruction_sequence *const sequence, unicode_view const name,
+                                               source_location const original_reference_location)
+{
+    LPG_FOR(size_t, i, state->local_variables.count)
+    {
+        local_variable const *const variable = state->local_variables.elements + i;
+        if (unicode_view_equals(unicode_view_from_string(variable->name), name))
+        {
+            return read_local_variable_result_create(
+                variable_address_from_local(variable->where), variable->type_, variable->compile_time_value, true);
+        }
+    }
+    if (!state->parent)
+    {
+        return read_local_variable_result_unknown;
+    }
+    read_local_variable_result const outer_variable =
+        read_local_variable(state->parent, NULL, name, original_reference_location);
+    switch (outer_variable.status)
+    {
+    case read_local_variable_status_ok:
+        break;
+
+    case read_local_variable_status_forbidden:
+    case read_local_variable_status_unknown:
+        return outer_variable;
+    }
+    if (outer_variable.compile_time_value.is_set && sequence)
+    {
+        register_id const where = allocate_register(&state->used_registers);
+        add_instruction(sequence, instruction_create_literal(literal_instruction_create(
+                                      where, outer_variable.compile_time_value.value_, outer_variable.what)));
+        return read_local_variable_result_create(variable_address_from_local(where), outer_variable.what,
+                                                 outer_variable.compile_time_value, outer_variable.is_pure);
+    }
+    if (!state->may_capture_runtime_variables)
+    {
+        state->on_error(
+            semantic_error_create(semantic_error_cannot_capture_runtime_variable, original_reference_location),
+            state->user);
+        return read_local_variable_result_forbidden;
+    }
+    capture_index const existing_capture = require_capture(state, outer_variable.where, outer_variable.what);
+    return read_local_variable_result_create(variable_address_from_capture(existing_capture), outer_variable.what,
+                                             outer_variable.compile_time_value, outer_variable.is_pure);
 }
